Moved group end points to a new groups modules, registered groups module in __init__.py
