+++ donut/__init__.py
@@ -15,6 +15,7 @@ from donut.modules import account
 from donut.modules import auth
 from donut.modules import marketplace
 from donut.modules import core
+from donut.modules import groups
 
 app = flask.Flask(__name__)
 
@@ -23,6 +24,7 @@ app.register_blueprint(account.blueprint)
 app.register_blueprint(auth.blueprint)
 app.register_blueprint(marketplace.blueprint)
 app.register_blueprint(core.blueprint)
+app.register_blueprint(groups.blueprint)
 
 
 def init(environment_name):
