+++ donut/modules/core/helpers.py
@@ -88,44 +88,6 @@ def get_member_list_data(fields=None, attrs={}):
     result = [{f: t for f, t in zip(fields, res)} for res in result]
     return result
 
-
-def get_group_list_data(fields=None, attrs={}):
-    """
-    Queries the database and returns list of group data constrained by the 
-    specified attributes.
-
-    Arguments:
-        fields: The fields to return. If None specified, then default_fields
-                are used.
-        attrs:  The attributes of the group to filter for.
-    Returns:
-        result: The fields and corresponding values of groups with desired
-                attributes. In the form of a list of dicts with key:value of 
-                columnname:columnvalue.
-    """
-    all_returnable_fields = ["group_id", "group_name", "group_desc"]
-    default_fields = ["group_id", "group_name", "group_desc"]
-    if fields == None:
-        fields = default_fields
-    else:
-        if any(f not in all_returnable_fields for f in fields):
-            return "Invalid field"
-
-    # Build the SELECT and FROM clauses
-    s = sqlalchemy.sql.select(fields).select_from(sqlalchemy.text("groups"))
-
-    # Build the WHERE clause
-    for key, value in list(attrs.items()):
-        s = s.where(sqlalchemy.text(key + "= :" + key))
-
-    # Execute the query
-    result = flask.g.db.execute(s, attrs).fetchall()
-
-    # Return the rows in the form of a list of dicts
-    result = [{f: t for f, t in zip(fields, res)} for res in result]
-    return result
-
-
 def get_name_and_email(user_id):
     """
     Queries the database and returns the full_name and email corresponding to the
