+++ donut/modules/groups/__init__.py
@@ -1,0 +1,4 @@
+import flask
+blueprint = flask.Blueprint('groups',__name__, template_folder='templates')
+
+import donut.modules.groups.routes
